prevent double

+++ boards/FSM/Src/Io/Io_WheelSpeedSensors.c
@@ -9,7 +9,7 @@ static const float  MPS_TO_KPH_CONVERSION_FACTOR = 3.6f;
 static const size_t RELUCTOR_RING_TOOTH_COUNT    = 48;
 static const float  TIRE_DIAMETER                = 0.4572f;
 static const float  ARC_LENGTH_PER_RELUCTOR_TOOTH =
-    (float)((M_PI * TIRE_DIAMETER) / RELUCTOR_RING_TOOTH_COUNT);
+    (float)(((float)M_PI * TIRE_DIAMETER) / RELUCTOR_RING_TOOTH_COUNT);
 
 static struct FreqOnlyPwmInput *left_wheel_speed_sensor,
     *right_wheel_speed_sensor;
